requirements.txt
flask
joblib
scikit-learn
xgboost
dnspython
requests